moved full node to its own file
